Fix vari

+++ cogs/ticket.py
@@ -11,32 +11,57 @@ class Ticket(commands.Cog):
 
     def __init__(self, bot):
         self.bot = bot
-        self.ticketChannel = 757654228259831872  # canale ticket
-        self.category = 757654228259831871  # categoria canale ticket
-        self.role = 757654227211518023  # support role
-        self.cache = 757654228259831873  # cache channels
+        self.ticketChannel = 748908550973292604  # canale ticket
+        self.category = 750086378028793908  # categoria canale ticket
+        self.role = 748907435174920283  # support role
+        self.cache = 762079923942195220  # cache channels
+        self.limited_roles = [754829854066737182, 744631301872680980] # limited roles 
+        self.limit = [2, 2] # number of tickets for limited roles
 
+    # => Help Command
     @commands.command()
-    async def ticket_help(self, ctx):
+    async def tickethelp(self, ctx):
+        cfg = self.bot.get_cog('Config') 
+        if cfg.rolea1 in [role.id for role in ctx.message.author.roles] or cfg.rolea2 in [role.id for role in ctx.message.author.roles]\
+        or cfg.rolea3 in [role.id for role in ctx.message.author.roles] or cfg.rolea4 in [role.id for role in ctx.message.author.roles]\
+        or cfg.rolea5 in [role.id for role in ctx.message.author.roles] or cfg.rolea6 in [role.id for role in ctx.message.author.roles]:
 
-        d = "!tAdd - Aggiungere un utente al ticket"
-        embed = discord.Embed(title="Comandi Ticket", description=d, color=discord.Colour.green())
-        await ctx.send(embed=embed)
-        await ctx.message.delete()
+            d = "it!tAdd - Aggiungere un utente al ticket"
+            embed = discord.Embed(title="Comandi Ticket", description=d, color=discord.Colour.green())
+            await ctx.send(embed=embed)
+            await ctx.message.delete()
 
+    # => Send Ticket Message
     @commands.command()
-    async def t_add(self, ctx, arg: discord.User):
+    async def tmessage(self, ctx):
+        cfg = self.bot.get_cog('Config') 
+        if cfg.rolea1 in [role.id for role in ctx.message.author.roles] or cfg.rolea2 in [role.id for role in ctx.message.author.roles]\
+        or cfg.rolea3 in [role.id for role in ctx.message.author.roles] or cfg.rolea4 in [role.id for role in ctx.message.author.roles]\
+        or cfg.rolea5 in [role.id for role in ctx.message.author.roles] or cfg.rolea6 in [role.id for role in ctx.message.author.roles]:
+            
+            embed = discord.Embed(title="Pannello Tickets", description = "Per aprire un ticket e contattare lo staff premi la reazione 🎫 sottostante.\nTi ricordo che qualsiasi ticket aperto e inutilizzato sarà frutto di warn.\n\nIn caso di problemi relativi ai sistemi sviluppati per Among Us Ita contatta un **Developers** o un **Mod**", color = discord.Colour.green())
+            z = await ctx.send(embed=embed)
+            await z.add_reaction('🎫')
+    
+    # => Add user to ticket
+    @commands.command()
+    async def tadd(self, ctx, arg: discord.User):
+        cfg = self.bot.get_cog('Config') 
+        if cfg.rolea1 in [role.id for role in ctx.message.author.roles] or cfg.rolea2 in [role.id for role in ctx.message.author.roles]\
+        or cfg.rolea3 in [role.id for role in ctx.message.author.roles] or cfg.rolea4 in [role.id for role in ctx.message.author.roles]\
+        or cfg.rolea5 in [role.id for role in ctx.message.author.roles] or cfg.rolea6 in [role.id for role in ctx.message.author.roles]:
+        
+            channel = ctx.message.channel
+            user_2_add = arg
 
-        channel = ctx.message.channel
-        user_2_add = arg
+            t = "Aggiunto utente"
+            d = f"L'utente {user_2_add.mention} è stato aggiunto alla stanza"
+            await ctx.message.delete()
+            await channel.set_permissions(user_2_add, read_messages=True, send_messages=True, add_reactions=False)
+            embed = discord.Embed(title=t, description=d)
+            await ctx.send(embed=embed)
 
-        t = "Aggiunto utente"
-        d = f"L'utente {user_2_add.mention} è stato aggiunto alla stanza"
-        await ctx.message.delete()
-        await channel.set_permissions(user_2_add, read_messages=True, send_messages=True, add_reactions=False)
-        embed = discord.Embed(title=t, description=d)
-        await ctx.send(embed=embed)
-
+    # => Listen reaction + Ticket Functions
     @commands.Cog.listener()
     async def on_raw_reaction_add(self, payload):
         user_id = payload.user_id
@@ -45,21 +70,24 @@ class Ticket(commands.Cog):
         emoji = payload.emoji.name
         guild_id = payload.guild_id
         member = payload.member
-
+        user_roles = [i.id for i in member.roles]
+        
         if user_id != self.bot.user.id:
             if channel_id == self.ticketChannel and emoji == '🎫':
+                
+                user = self.bot.get_user(user_id)
                 channel = self.bot.get_channel(self.ticketChannel)
                 msg = await channel.fetch_message(message_id)
-                user = self.bot.get_user(user_id)
+                
                 await msg.remove_reaction('🎫', user)
                 await Ticket.create_channel(self, user_id, guild_id)
 
             channel = self.bot.get_channel(channel_id)
-            is_support = True if self.role in [i.id for i in member.roles] else False
+            is_support = True if self.role in user_roles else False
             if 'ticket-' in channel.name and is_support:
-                if emoji == '✅':
+                if emoji == '✅':                    
                     # Claim
-                    await Ticket.claim_ticket(self, user_id, message_id, channel_id, guild_id)
+                    await Ticket.claim_ticket(self, user_id, message_id, channel_id, guild_id, user_roles)
                 elif emoji == '🟡':
                     # Non è stato possibile
                     message = ("Mi dispiace", "Non è stato possibile chiudere il ticket in maniera corretta", 15844367)
@@ -114,31 +142,44 @@ class Ticket(commands.Cog):
 
         await conn.commit()
 
-    async def claim_ticket(self, user_id, message_id, channel_id, guild_id):
+    async def claim_ticket(self, user_id, message_id, channel_id, guild_id, user_roles):
 
         conn = self.bot.get_cog('Db')
-        user_ = conn.fetchall("SELECT * FROM tickets WHERE channel_id = ?", (channel_id,))[0][1]
+        
+        check_n_ticket = conn.fetchall("SELECT * FROM tickets WHERE admin_id = ?", (user_id,))
 
-        await conn.commit()
-
+        admin = self.bot.get_user(user_id)
         channel = self.bot.get_channel(channel_id)
-
         guild = self.bot.get_guild(guild_id)
         supporter = guild.get_role(self.role)
-        user_ = self.bot.get_user(user_)
-        user = self.bot.get_user(user_id)
+        
+        utente = conn.fetchall("SELECT * FROM tickets WHERE channel_id = ?", (channel_id,))[0][1]
+        utente = self.bot.get_user(utente)
+        
+        m = await channel.fetch_message(message_id)
+        
+        if len(check_n_ticket) > 0:
+            for idx, i in enumerate(self.limited_roles):
+                if i in user_roles and len(check_n_ticket) == self.limit[idx]:
+                    await m.remove_reaction('✅', admin)
+                    embed = discord.Embed(title="Errore", description="Mi dispiace ma hai raggiunto il numero massimo di ticket claimabili, per claimare questo ticket chiudo un tuo ticket.")
+                    await admin.send(embed=embed)
+                    return 0
+                        
+        conn.execute("UPDATE tickets SET admin_id = ? WHERE channel_id = ?", (user_id, channel_id,))
+        
+        await conn.commit()        
 
         await channel.set_permissions(supporter, read_messages=False, send_messages=False, add_reactions=False)
-        await channel.set_permissions(user_, read_messages=True, send_messages=True, add_reactions=False)
-        await channel.set_permissions(user, read_messages=True, send_messages=True, add_reactions=True)
+        await channel.set_permissions(utente, read_messages=True, send_messages=True, add_reactions=False)
+        await channel.set_permissions(admin, read_messages=True, send_messages=True, add_reactions=True)
 
-        m = await channel.fetch_message(message_id)
         await m.clear_reactions()
         await m.delete()
 
         embed = discord.Embed(title=m.embeds[0].title,
-                              description=f"Ciao {user_.mention} la tua richiesta è stata presa in carico dallo "
-                                          f"Staffer {user.mention} di Among Us Ita, come può esserti utile?\n"
+                              description=f"Ciao {utente.mention} la tua richiesta è stata presa in carico dallo "
+                                          f"Staffer {admin.mention} di Among Us Ita, come può esserti utile?\n"
                                           f"Esponi chiaramente la tua richiesta affinchè lo staffer possa "
                                           f"esaustivamente risolvere la tua problematica, ricordati che l'apertura di "
                                           f"ticket inutilizzati incomberà al warn.")
@@ -148,10 +189,10 @@ class Ticket(commands.Cog):
                               "🟡 - Impossibile risolvere il ticket\n"
                               "🔴 - Ticket risolto correttamente")
 
-        m = await channel.send(embed=embed)
-        await m.add_reaction('🔵')
-        await m.add_reaction('🟡')
-        await m.add_reaction('🔴')
+        msg = await channel.send(embed=embed)
+        await msg.add_reaction('🔵')
+        await msg.add_reaction('🟡')
+        await msg.add_reaction('🔴')
 
     async def delete_channel(self, channel_id):
 
@@ -171,8 +212,9 @@ class Ticket(commands.Cog):
     async def cache_messages(self, channel_id):
 
         conn = self.bot.get_cog('Db')
-        n_ticket = conn.fetchall("SELECT * FROM tickets WHERE channel_id = ?", (channel_id,))[0][0]
+        ticket = conn.fetchall("SELECT * FROM tickets WHERE channel_id = ?", (channel_id,))
 
+        n_ticket = ticket[0][0]
         channel = self.bot.get_channel(channel_id)
         cache = self.bot.get_channel(self.cache)
 
@@ -186,14 +228,26 @@ class Ticket(commands.Cog):
                 text = f"{author.name}: {msg}"
                 cached_messages.append(text)
 
+        title = f"Ticket N: {n_ticket}"
+        
+        user = self.bot.get_user(ticket[0][1])
+        admin = self.bot.get_user(ticket[0][3])
+        
         if len(cached_messages) == 0:
-            embed = discord.Embed(title=f"Ticket N: {n_ticket}", description="Nessun Messaggio")
+            description = "Nessun Messaggio"
+            embed = discord.Embed(title = title, description = description)
+            embed.add_field(name="Creatore", value=f"{user.mention}")
+            embed.add_field(name="Admin", value=f"{admin.mention}")
             await cache.send(embed=embed)
         else:
             with open(f'{n_ticket}.txt', 'w+') as f:
                 f.write('\n'.join(cached_messages))
-            embed = discord.Embed(title=f"Ticket N: {n_ticket}")
+                
+            embed = discord.Embed(title = title)
+            embed.add_field(name="Creatore", value=f"{user.mention}")
+            embed.add_field(name="Admin", value=f"{admin.mention}")
             await cache.send(embed=embed)
+            
             with open(f'{n_ticket}.txt', 'rb') as f:
                 await cache.send(file=discord.File(f))
